Enabling in-memory DataContextV3 instantiation.
